+++ create-gd-bucket-and-key/modules/s3-bucket-create/providers.tf
@@ -19,3 +19,7 @@
 provider "aws" {
   alias = "src"
 }
+
+provider "aws" {
+  alias = "key"
+}