+++ import-org/org.tf
@@ -1,3 +1,4 @@
+
 #  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #  SPDX-License-Identifier: MIT-0
 
@@ -19,10 +20,5 @@ resource "aws_organizations_organization" "my_org" {
 
   feature_set = "ALL"
 
-  aws_service_access_principals = [
-    "guardduty.amazonaws.com",
-    "config-multiaccountsetup.amazonaws.com",
-    "securityhub.amazonaws.com",
-    "config.amazonaws.com"
-  ]
+  aws_service_access_principals = ["sso.amazonaws.com", "config.amazonaws.com", "securityhub.amazonaws.com", "guardduty.amazonaws.com", "config-multiaccountsetup.amazonaws.com"]
 }