+++ scripts/full-setup.sh
@@ -34,7 +34,7 @@ export TF_VAR_role_to_assume_for_role_creation=`cat configuration.json | jq -r "
 export TF_VAR_finding_publishing_frequency=`cat configuration.json | jq -r ".finding_publishing_frequency"`
 export TF_VAR_guardduty_findings_bucket_region=`cat configuration.json | jq -r ".guardduty_findings_bucket_region"`
 export TF_VAR_logging_acc_s3_bucket_name=`cat configuration.json | jq -r ".logging_acc_s3_bucket_name"`
-export TF_VAR_logging_acc_kms_key_alias=`cat configuration.json | jq -r ".logging_acc_kms_key_alias"`
+export TF_VAR_security_acc_kms_key_alias=`cat configuration.json | jq -r ".security_acc_kms_key_alias"`
 export TF_VAR_s3_access_log_bucket_name=`cat configuration.json | jq -r ".s3_access_log_bucket_name"`
 
 # Set default region of operation